% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/methods.R
\docType{methods}
\name{brp,FLBRP-method}
\alias{brp,FLBRP-method}
\title{Fit an FLBRP object}
\usage{
\S4method{brp}{FLBRP}(object)
}
\arguments{
\item{object}{An object of class \code{FLBRP}}
}
\value{
An object of class \code{FLBRP} with reference points estimated
}
\description{
This method carries out the necessary calculation to obtain the reference
points for an object of class \code{FLBRP}. Results of the calculation are stored
in \code{@refpts} slot of the object, and the object is then returned.
}
\examples{
data(ple4brp)

ple4brp <- brp(ple4brp)
refpts(ple4brp)
}
\seealso{
\link{FLBRP}
}
