% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/classes.R
\docType{class}
\name{FLBRP}
\alias{FLBRP}
\alias{FLBRP-methods}
\alias{FLBRP-class}
\title{The FLR class for biological and economic reference points}
\description{
Lorem ipsum dolor sit amet, consectetur adipiscing elit. Pellentesque eleifend
odio ac rutrum luctus. Aenean placerat porttitor commodo. Pellentesque eget porta
libero. Pellentesque molestie mi sed orci feugiat, non mollis enim tristique. 
Suspendisse eu sapien vitae arcu lobortis ultrices vitae ac velit. Curabitur id
}
\section{Slots}{

\describe{
\item{\code{model}}{The formula for the stock-recruitment model, `formula,`.}

\item{\code{params}}{Parameters of the stock-recruitment relationship, `FLPar,`.}

\item{\code{refpts}}{The estimated reference points, `FLPar,`.}

\item{\code{fbar}}{A vector of mean fishing mortality values for calculating the reference points, `FLQuant,`.}

\item{\code{fbar.obs}}{The time series of observed mean fishing mortality, `FLQuant,`.}

\item{\code{landings.obs}}{The time series of observed total landings, `FLQuant,`.}

\item{\code{discards.obs}}{The time series of observed total discards, `FLQuant,`.}

\item{\code{rec.obs}}{The time series of observed recruitment, `FLQuant,`.}

\item{\code{ssb.obs}}{The time series of observed spawning stock biomasses, `FLQuant,`.}

\item{\code{stock.obs}}{The time series of observed total stock abundance, `FLQuant,`.}

\item{\code{profit.obs}}{The time series of profits, `FLQuant,`.}

\item{\code{landings.sel}}{Selectivity at age for the landings, `FLQuant,`.}

\item{\code{discards.sel}}{Selectivity at age for the discards, `FLQuant,`.}

\item{\code{bycatch.harvest}}{X, Fishing mortality at age of bycatch species, `FLQuant,`.}

\item{\code{stock.wt}}{Mean weight-at-age in the stock to use in calculations, `FLQuant,`.}

\item{\code{landings.wt}}{Mean weight-at-age in the landings to use in calculations, `FLQuant,`.}

\item{\code{discards.wt}}{Mean weight-at-age in the discards to use in calculations, `FLQuant,`.}

\item{\code{bycatch.wt}}{Mean weight-at-age in the bycatch to use in calculations, `FLQuant,`.}

\item{\code{m}}{Natural mortality at age, `FLQuant,`.}

\item{\code{mat}}{Maturity at age, `FLQuant,`.}

\item{\code{harvest.spwn}}{Proportion of harvest/fishing mortality before spawning, `FLQuant,`.}

\item{\code{m.spwn}}{Proportion of natural mortality before spawning, `FLQuant,`.}

\item{\code{availability}}{, `FLQuant,`.}

\item{\code{price}}{Price at age, `FLQuant,`.}

\item{\code{vcost}}{X, `FLQuant,`.}

\item{\code{fcost}}{X, `FLQuant,`.}
}}

\section{Validity}{


  \describe{
    \item{VALIDITY}{Neque porro quisquam est qui dolorem ipsum.}
}

You can inspect the class validity function by using
   \code{getValidity(getClassDef('FLBRP'))}
}

\section{Accessors}{

All slots in the class have accessor and replacement methods defined that
allow retrieving and substituting individual slots.

The values passed for replacement need to be of the class of that slot.
A numeric vector can also be used when replacing FLQuant slots, and the
vector will be used to substitute the values in the slot, but not its other
attributes.
}

\section{Constructor}{

A construction method exists for this class that can take named arguments for
any of its slots. All slots are then created to match the requirements of the
class validity. If an unnamed \code{FLQuant} object is provided, this is used
for sizing but not stored in any slot.

A standard method also exists for object of class \code{FLStock}, with or
without additional information on the stock-recruits relationship. This can
be provided as an object of class \code{FLSR}, one of class \code{predictModel}
or as a list wqith elements named 'params', of class \code{FLPar}, and 'model',
of class \code{formula}.
}

\section{Methods}{

Methods exist for various calculations based on values stored in the class:

\describe{
    \item{METHOD}{Neque porro quisquam est qui dolorem ipsum.}
}
}

\examples{

}
\seealso{
\link{FLComp} \link{FLStock} \link{FLSR} \link{brp}
}
\author{
The FLR Team
}
\keyword{classes}
